x/eywacontract/types/message_handshake_test.go
package types

import (
	"testing"

	"eywa-contract/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgHandshake_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgHandshake
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgHandshake{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgHandshake{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
